refactor: 패키지 구조 변경에 따른 import 문 업데이트
- auth 모듈: application.auth → auth.application으로 패키지 재구조화
- schedule 모듈: application.schedule → schedule.application으로 패키지 재구조화
- 모든 import 문 경로를 새로운 계층형 아키텍처에 맞게 업데이트
- 테스트 코드 import 경로도 함께 수정
