+++ src/main/java/com/better/CommuteMate/auth/controller/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.better.CommuteMate.controller.auth.dto;
+package com.better.CommuteMate.auth.controller.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
