+++ src/main/java/com/better/CommuteMate/global/controller/GlobalExceptionHandler.java
@@ -1,13 +1,12 @@
 package com.better.CommuteMate.global.controller;
 
-import com.better.CommuteMate.application.schedule.exceptions.ScheduleAllFailureException;
-import com.better.CommuteMate.application.schedule.exceptions.SchedulePartialFailureException;
+import com.better.CommuteMate.schedule.application.exceptions.ScheduleAllFailureException;
+import com.better.CommuteMate.schedule.application.exceptions.SchedulePartialFailureException;
 import com.better.CommuteMate.global.controller.dtos.Response;
 import com.better.CommuteMate.global.exceptions.BasicException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
