+++ src/main/java/com/better/CommuteMate/schedule/application/exceptions/ScheduleAllFailureException.java
@@ -1,4 +1,4 @@
-package com.better.CommuteMate.application.schedule.exceptions;
+package com.better.CommuteMate.schedule.application.exceptions;
 
 import com.better.CommuteMate.global.controller.dtos.ErrorResponseDetail;
 import com.better.CommuteMate.global.exceptions.BasicException;
