+++ src/main/java/com/better/CommuteMate/schedule/controller/schedule/dtos/WorkScheduleDTO.java
@@ -1,6 +1,6 @@
-package com.better.CommuteMate.controller.schedule.dtos;
+package com.better.CommuteMate.schedule.controller.schedule.dtos;
 
-import com.better.CommuteMate.application.schedule.dtos.WorkScheduleCommand;
+import com.better.CommuteMate.schedule.application.dtos.WorkScheduleCommand;
 
 import java.time.LocalDateTime;
